+++ examples/blog-next/app/sitemap.ts
@@ -1,0 +1,1 @@
+export {generateSitemap as default} from "@/data/sanity/client";