+++ examples/blog-studio/src/schemas/documents/index.ts
@@ -1,0 +1,19 @@
+import blogAuthor from "./blog-author";
+import blogIndex from "./blog-index";
+import blogPost from "./blog-post";
+import blogTag from "./blog-tag";
+import home from "./home";
+import modularPage from "./modular.page";
+import settings from "./settings";
+
+const documents = [
+	home,
+	modularPage,
+	blogIndex,
+	blogPost,
+	blogTag,
+	blogAuthor,
+	settings,
+];
+
+export default documents;