+++ packages/sanity-studio/src/utils/index.ts
@@ -1,3 +1,10 @@
+export {
+	definePortableTextFactory,
+	defaultPTDecorators,
+	defaultPTLists,
+	defaultPTStyles,
+} from "./create-portable-text";
 export {definePathname} from "./definePathname";
+export {isUnique} from "./is-unique";
 export {localizedItem} from "./localizedItem";
 export {singletonListItem} from "./singleton-list-item";