+++ apps/next/CHANGELOG.md
@@ -1,5 +1,18 @@
 # next-non-embedded-studio
 
+## 1.3.10
+
+### Patch Changes
+
+- Updated dependencies [cc1b86d]
+- Updated dependencies [947c643]
+- Updated dependencies [947c643]
+- Updated dependencies [947c643]
+  - @tinloof/sanity-next@1.3.0
+  - @tinloof/sanity-studio@1.16.0
+  - @tinloof/sanity-web@1.1.0
+  - @examples/hello-world-i18n-studio@0.0.8
+
 ## 1.3.9
 
 ### Patch Changes